requirements.txt
# AI Multi-Agent Coding Assistant - Dependencies

# Core AutoGen packages
autogen-agentchat>=0.4.7
autogen-ext[gemini]>=0.4.7

# Environment variables
python-dotenv>=1.0.0

# Web interface
streamlit>=1.50.0

# Additional dependencies (installed automatically with above)
# google-genai
# tenacity
# websockets
# pandas
# numpy
# altair